samples/python/verify-winml/requirements.txt
foundry-local-sdk-winml
openai
